rename “configSettingDefinitions” to “settingDefinitions”
